+++ pom.xml
@@ -58,8 +58,8 @@
 		</dependency>
 		<dependency>
 			<groupId>io.github.oneteme</groupId>
-			<artifactId>jquery</artifactId>
-			<version>5.0.0-SNAPSHOT</version>
+			<artifactId>jquery-spring-boot-starter</artifactId>
+			<version>0.0.1-SNAPSHOT</version>
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
