+++ src/main/java/org/usf/inspect/server/dao/PurgeDao.java
@@ -11,8 +11,8 @@ import org.usf.inspect.core.InspectCollectorConfiguration;
 import org.usf.inspect.core.InstanceEnvironment;
 import org.usf.inspect.core.InstanceType;
 import org.usf.inspect.core.RestRemoteServerProperties;
-import org.usf.inspect.server.repo.InspectStore;
-import org.usf.inspect.server.repo.InstanceCatalog;
+import org.usf.inspect.server.erm.InspectStore;
+import org.usf.inspect.server.erm.InstanceCatalog;
 import org.usf.jquery.core.Column;
 import org.usf.jquery.core.Order;
 import org.usf.jquery.mvc.StoreManager;
