+++ src/main/java/org/usf/inspect/server/erm/DatabaseStageCatalog.java
@@ -1,11 +1,12 @@
-package org.usf.inspect.server.repo;
+package org.usf.inspect.server.erm;
 
-import static org.usf.inspect.server.config.constant.FieldConstant.CD_LDAP_RQT;
+import static org.usf.inspect.server.config.constant.FieldConstant.CD_DTB_RQT;
 import static org.usf.inspect.server.config.constant.FieldConstant.CD_ORD;
 import static org.usf.inspect.server.config.constant.FieldConstant.DH_END;
 import static org.usf.inspect.server.config.constant.FieldConstant.DH_STR;
 import static org.usf.inspect.server.config.constant.FieldConstant.VA_ARG;
 import static org.usf.inspect.server.config.constant.FieldConstant.VA_CMD;
+import static org.usf.inspect.server.config.constant.FieldConstant.VA_CNT;
 import static org.usf.inspect.server.config.constant.FieldConstant.VA_NAM;
 
 import org.usf.jquery.core.Column;
@@ -13,7 +14,7 @@ import org.usf.jquery.core.ViewColumn;
 import org.usf.jquery.mvc.Bind;
 import org.usf.jquery.mvc.DatasetCatalog;
 
-public interface LdapStageCatalog extends StageCatalog {
+public interface DatabaseStageCatalog extends StageCatalog {
 	
 	@Bind(DH_STR)
 	ViewColumn start();
@@ -21,15 +22,17 @@ public interface LdapStageCatalog extends StageCatalog {
 	@Bind(DH_END)
 	ViewColumn end();
 	
+	@Bind(VA_CNT)
+	ViewColumn actionCount();
+	
 	@Bind(VA_ARG)
 	ViewColumn arg();
 	
 	@Bind(VA_CMD)
 	ViewColumn command();
 	
-	@Bind(CD_LDAP_RQT)
+	@Bind(CD_DTB_RQT)
 	ViewColumn parent();
-
 	
 	default Column elapsedTime() {
 		return end().minus(start()).epoch();
