+++ src/main/java/org/usf/inspect/server/erm/FtpStageCatalog.java
@@ -1,9 +1,10 @@
-package org.usf.inspect.server.repo;
+package org.usf.inspect.server.erm;
 
+import static org.usf.inspect.server.config.constant.FieldConstant.CD_FTP_RQT;
 import static org.usf.inspect.server.config.constant.FieldConstant.CD_ORD;
-import static org.usf.inspect.server.config.constant.FieldConstant.CD_SMTP_RQT;
 import static org.usf.inspect.server.config.constant.FieldConstant.DH_END;
 import static org.usf.inspect.server.config.constant.FieldConstant.DH_STR;
+import static org.usf.inspect.server.config.constant.FieldConstant.VA_ARG;
 import static org.usf.inspect.server.config.constant.FieldConstant.VA_CMD;
 import static org.usf.inspect.server.config.constant.FieldConstant.VA_NAM;
 
@@ -12,7 +13,7 @@ import org.usf.jquery.core.ViewColumn;
 import org.usf.jquery.mvc.Bind;
 import org.usf.jquery.mvc.DatasetCatalog;
 
-public interface SmtpStageCatalog extends StageCatalog {
+public interface FtpStageCatalog extends StageCatalog {
 	
 	@Bind(DH_STR)
 	ViewColumn start();
@@ -20,10 +21,13 @@ public interface SmtpStageCatalog extends StageCatalog {
 	@Bind(DH_END)
 	ViewColumn end();
 	
+	@Bind(VA_ARG)
+	ViewColumn arg();
+	
 	@Bind(VA_CMD)
 	ViewColumn command();
 	
-	@Bind(CD_SMTP_RQT)
+	@Bind(CD_FTP_RQT)
 	ViewColumn parent();
 	
 	default Column elapsedTime() {
