+++ src/main/java/org/usf/inspect/server/erm/RequestCatalog.java
@@ -1,0 +1,64 @@
+package org.usf.inspect.server.erm;
+
+import static org.usf.inspect.server.config.constant.FieldConstant.CD_INS;
+import static org.usf.inspect.server.config.constant.FieldConstant.CD_PRT;
+import static org.usf.inspect.server.config.constant.FieldConstant.DH_END;
+import static org.usf.inspect.server.config.constant.FieldConstant.DH_STR;
+import static org.usf.inspect.server.config.constant.FieldConstant.VA_HST;
+import static org.usf.inspect.server.config.constant.FieldConstant.VA_THR;
+import static org.usf.inspect.server.config.constant.FieldConstant.VA_USR;
+import static org.usf.jquery.core.Join.innerJoin;
+import static org.usf.jquery.core.Join.leftJoin;
+import static org.usf.jquery.core.JoinGroup.joins;
+import static org.usf.jquery.mvc.StoreManager.getInstance;
+
+import org.usf.inspect.core.RequestMask;
+import org.usf.jquery.core.Column;
+import org.usf.jquery.core.JoinGroup;
+import org.usf.jquery.core.ViewColumn;
+import org.usf.jquery.mvc.Bind;
+import org.usf.jquery.mvc.DatasetCatalog;
+import org.usf.jquery.mvc.Expose;
+
+public interface RequestCatalog extends DatasetCatalog {
+
+    ViewColumn id();
+
+    @Bind(VA_HST)
+    ViewColumn host();
+
+    @Bind(CD_PRT)
+    ViewColumn port();
+
+    @Bind(DH_STR)
+    ViewColumn start();
+
+    @Bind(DH_END)
+    ViewColumn end();
+
+    @Bind(VA_USR)
+    ViewColumn user();
+
+    @Bind(VA_THR)
+    ViewColumn thread();
+
+    @Bind(CD_INS)
+    @Expose(identity = "instance_env")
+    ViewColumn instanceEnv();
+	
+	default Column elapsedTime() {
+		return end().minus(start()).epoch();
+	}
+
+    default JoinGroup instance() {
+        var instance = getInstance().getStore(InspectStore.class).instance();
+        return joins(innerJoin(instance.getView(), instanceEnv().eq(instance.id())));
+    }
+	
+	default JoinGroup exception() {
+		var exception = getInstance().getStore(InspectStore.class).exception();
+		return joins(leftJoin(exception.getView(), id().eq(exception.parent()), exception.type().eq(getRequestType().name())));
+	}
+    
+    RequestMask getRequestType();
+}
