+++ src/main/java/org/usf/inspect/server/erm/RestRequestStageCatalog.java
@@ -1,12 +1,9 @@
-package org.usf.inspect.server.repo;
+package org.usf.inspect.server.erm;
 
-import static org.usf.inspect.server.config.constant.FieldConstant.CD_DTB_RQT;
 import static org.usf.inspect.server.config.constant.FieldConstant.CD_ORD;
+import static org.usf.inspect.server.config.constant.FieldConstant.CD_PRN_SES;
 import static org.usf.inspect.server.config.constant.FieldConstant.DH_END;
 import static org.usf.inspect.server.config.constant.FieldConstant.DH_STR;
-import static org.usf.inspect.server.config.constant.FieldConstant.VA_ARG;
-import static org.usf.inspect.server.config.constant.FieldConstant.VA_CMD;
-import static org.usf.inspect.server.config.constant.FieldConstant.VA_CNT;
 import static org.usf.inspect.server.config.constant.FieldConstant.VA_NAM;
 
 import org.usf.jquery.core.Column;
@@ -14,7 +11,7 @@ import org.usf.jquery.core.ViewColumn;
 import org.usf.jquery.mvc.Bind;
 import org.usf.jquery.mvc.DatasetCatalog;
 
-public interface DatabaseStageCatalog extends StageCatalog {
+public interface RestRequestStageCatalog extends StageCatalog {
 	
 	@Bind(DH_STR)
 	ViewColumn start();
@@ -22,16 +19,7 @@ public interface DatabaseStageCatalog extends StageCatalog {
 	@Bind(DH_END)
 	ViewColumn end();
 	
-	@Bind(VA_CNT)
-	ViewColumn actionCount();
-	
-	@Bind(VA_ARG)
-	ViewColumn arg();
-	
-	@Bind(VA_CMD)
-	ViewColumn command();
-	
-	@Bind(CD_DTB_RQT)
+	@Bind(CD_PRN_SES)
 	ViewColumn parent();
 	
 	default Column elapsedTime() {
