--- src/main/java/org/usf/inspect/server/repo/SmtpRequestCatalog.java
@@ -1,42 +1,0 @@
-package org.usf.inspect.server.repo;
-
-import static org.usf.inspect.core.RequestMask.SMTP;
-import static org.usf.inspect.server.config.constant.FieldConstant.CD_PRN_SES;
-import static org.usf.inspect.server.config.constant.FieldConstant.CD_PRT;
-import static org.usf.inspect.server.config.constant.FieldConstant.ID_SMTP_RQT;
-import static org.usf.inspect.server.config.constant.FieldConstant.VA_CMD;
-import static org.usf.inspect.server.config.constant.FieldConstant.VA_FAIL;
-import static org.usf.inspect.server.config.constant.FieldConstant.VA_THR;
-import static org.usf.inspect.server.config.constant.FieldConstant.VA_USR;
-import static org.usf.jquery.core.Join.leftJoin;
-import static org.usf.jquery.core.JoinGroup.joins;
-import static org.usf.jquery.mvc.StoreManager.getInstance;
-
-import org.usf.jquery.core.Column;
-import org.usf.jquery.core.JoinGroup;
-import org.usf.jquery.core.ViewColumn;
-import org.usf.jquery.mvc.Bind;
-
-public interface SmtpRequestCatalog extends RequestCatalog {
-
-	@Bind(ID_SMTP_RQT)
-	ViewColumn id();
-	
-	@Bind(VA_CMD)
-	ViewColumn command();
-	
-	@Bind(VA_FAIL)
-	ViewColumn failed();
-	
-	@Bind(CD_PRN_SES)
-	ViewColumn parent();
-	
-	default Column elapsedTime() {
-		return end().minus(start()).epoch();
-	}
-	
-	default JoinGroup exception() {
-		var exception = getInstance().getStore(InspectStore.class).exception();
-		return joins(leftJoin(exception.getView(), exception.parent().eq(id()), exception.type().eq(SMTP.name())));
-	}
-}
